+++ packages/components/molecules/f-navigation-links/src/components/_tests/NavigationLinks.test.js
@@ -68,7 +68,7 @@ describe('NavigationLinks', () => {
         expect(links).toMatchSnapshot();
     });
 
-    it('should pass href attribute to `<v-link>` when present', () => {
+    it('should pass to attribute to `<v-link>` when present', () => {
         const expected = '/account/info';
         const wrapper = shallowMount(NavigationLinks, {
             propsData: {
@@ -123,4 +123,59 @@ describe('NavigationLinks', () => {
         expect(link.attributes('href')).toEqual(expected);
         expect(link.attributes('to')).toBeUndefined();
     });
+
+    it('should skip link item if both `to` and `href` missing in props.links object', () => {
+        const wrapper = shallowMount(NavigationLinks, {
+            propsData: {
+                links: [
+                    {
+                        id: 'link1',
+                        name: 'Your account'
+                    },
+                    {
+                        id: 'link2',
+                        to: '/account/info',
+                        name: 'Your account'
+                    }]
+            }
+        });
+
+        const links = wrapper.find('[data-test-id="navigationLinks"]');
+        const link = wrapper.find('[data-test-id="link1"]');
+
+        expect(links.findAll('li').length).toEqual(1);
+        expect(link.exists()).toBeFalsy();
+    });
+
+    describe('computed ::', () => {
+        describe('filterLinks ::', () => {
+            it('should filter props.links if link does not have `to` or `href` property', () => {
+                const wrapper = shallowMount(NavigationLinks, {
+                    propsData: {
+                        links: [
+                            {
+                                id: 'link1',
+                                name: 'Link 1'
+                            },
+                            {
+                                id: 'link2',
+                                to: '/account/info',
+                                name: 'Your account'
+                            },
+                            {
+                                id: 'link3',
+                                name: 'Link 3',
+                                href: '/link-3'
+                            }]
+                    }
+                });
+
+                const result = wrapper.vm.filterLinks;
+                expect(result.length).toEqual(2);
+                expect(result.find(x => x.id === 'link1')).toBeUndefined();
+            });
+        });
+    });
 });
+
+